Fix Dxyn (Display instruction)
I don't know why it didn't work before, but it didn't. Also this might be
faster? I don't plan on profiling this pile of crap. I need to make it set
VF it it changes any pixel but as far as I know PONG doesn't use it so
I'll see you later sucker.